+++ src/libs/actions/IOU/index.ts
@@ -38,13 +38,12 @@ import {getManagerMcTestParticipant} from '@libs/OptionsListUtils';
 import {getCustomUnitID} from '@libs/PerDiemRequestUtils';
 import {addSMSDomainIfPhoneNumber} from '@libs/PhoneNumber';
 import {getDistanceRateCustomUnit, hasDependentTags, isPaidGroupPolicy} from '@libs/PolicyUtils';
-import {getAllReportActions, getOriginalMessage, getReportActionHtml, getReportActionText, isMoneyRequestAction, isReportPreviewAction} from '@libs/ReportActionsUtils';
+import {getOriginalMessage, getReportActionHtml, getReportActionText, isReportPreviewAction} from '@libs/ReportActionsUtils';
 import type {OptimisticChatReport, OptimisticCreatedReportAction, OptimisticIOUReportAction, TransactionDetails} from '@libs/ReportUtils';
 import {
     buildOptimisticAddCommentReportAction,
     buildOptimisticChatReport,
     buildOptimisticCreatedReportAction,
-    buildOptimisticCreatedReportForUnapprovedAction,
     buildOptimisticExpenseReport,
     buildOptimisticIOUReport,
     buildOptimisticIOUReportAction,
@@ -71,7 +70,6 @@ import {
     isOneTransactionThread,
     isOptimisticPersonalDetail,
     isPolicyExpenseChat as isPolicyExpenseChatReportUtil,
-    isProcessingReport,
     isSelectedManagerMcTest,
     isSelfDM,
     isTestTransactionReport,
@@ -4337,438 +4335,6 @@ function createDistanceRequest(distanceRequestInformation: CreateDistanceRequest
     return {iouReport: distanceIouReport};
 }
 
-type OptimisticHoldReportExpenseActionID = {
-    optimisticReportActionID: string;
-    oldReportActionID: string;
-};
-
-function getHoldReportActionsAndTransactions(reportID: string | undefined) {
-    const iouReportActions = getAllReportActions(reportID);
-    const holdReportActions: Array<OnyxTypes.ReportAction<typeof CONST.REPORT.ACTIONS.TYPE.IOU>> = [];
-    const holdTransactions: OnyxTypes.Transaction[] = [];
-
-    for (const action of Object.values(iouReportActions)) {
-        const transactionID = isMoneyRequestAction(action) ? getOriginalMessage(action)?.IOUTransactionID : undefined;
-        const transaction = allTransactions?.[`${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`];
-
-        if (transaction?.comment?.hold) {
-            holdReportActions.push(action as OnyxTypes.ReportAction<typeof CONST.REPORT.ACTIONS.TYPE.IOU>);
-            holdTransactions.push(transaction);
-        }
-    }
-
-    return {holdReportActions, holdTransactions};
-}
-
-type OptimisticReportActionCopyIDs = Record<string, string>;
-
-/**
- * Gets duplicate workflow actions for a partial expense report.
- * Used when splitting held expenses into a new partial report to maintain action history.
- *
- * @param sourceReportID - The ID of the original report to copy actions from
- * @param targetReportID - The ID of the new partial expense report to copy actions to
- * @returns A tuple of [optimisticData, successData, failureData, duplicatedReportActionIDs]
- */
-function getDuplicateActionsForPartialReport(
-    sourceReportID: string | undefined,
-    targetReportID: string | undefined,
-): [
-    Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>>,
-    Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>>,
-    Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>>,
-    OptimisticReportActionCopyIDs,
-] {
-    const optimisticData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>> = [];
-    const successData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>> = [];
-    const failureData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>> = [];
-    const optimisticReportActionCopyIDs: OptimisticReportActionCopyIDs = {};
-
-    if (!sourceReportID || !targetReportID) {
-        return [optimisticData, successData, failureData, optimisticReportActionCopyIDs];
-    }
-
-    const sourceReportActions = getAllReportActions(sourceReportID);
-
-    // Match the backend's WORKFLOW_ACTIONS list
-    const workflowActionTypes = [
-        CONST.REPORT.ACTIONS.TYPE.SUBMITTED,
-        CONST.REPORT.ACTIONS.TYPE.SUBMITTED_AND_CLOSED,
-        CONST.REPORT.ACTIONS.TYPE.APPROVED,
-        CONST.REPORT.ACTIONS.TYPE.UNAPPROVED,
-        CONST.REPORT.ACTIONS.TYPE.REJECTED,
-        CONST.REPORT.ACTIONS.TYPE.REJECTED_TO_SUBMITTER,
-        CONST.REPORT.ACTIONS.TYPE.RETRACTED,
-        CONST.REPORT.ACTIONS.TYPE.CLOSED,
-        CONST.REPORT.ACTIONS.TYPE.REOPENED,
-        CONST.REPORT.ACTIONS.TYPE.FORWARDED,
-        CONST.REPORT.ACTIONS.TYPE.TAKE_CONTROL,
-        CONST.REPORT.ACTIONS.TYPE.REROUTE,
-    ] as const;
-
-    const copiedActions: Record<string, OnyxTypes.ReportAction> = {};
-    const copiedActionsSuccess: OnyxCollection<NullishDeep<ReportAction>> = {};
-    const copiedActionsFailure: Record<string, null> = {};
-
-    for (const action of Object.values(sourceReportActions)) {
-        if (action && (workflowActionTypes as readonly string[]).includes(action.actionName)) {
-            const newActionID = NumberUtils.rand64();
-            copiedActions[newActionID] = {
-                ...action,
-                reportActionID: newActionID,
-                reportID: targetReportID,
-                pendingAction: CONST.RED_BRICK_ROAD_PENDING_ACTION.ADD,
-            };
-            copiedActionsSuccess[newActionID] = {
-                pendingAction: null,
-            };
-            copiedActionsFailure[newActionID] = null;
-            optimisticReportActionCopyIDs[action.reportActionID] = newActionID;
-        }
-    }
-
-    if (Object.keys(copiedActions).length > 0) {
-        optimisticData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${targetReportID}`,
-            value: copiedActions,
-        });
-
-        successData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${targetReportID}`,
-            value: copiedActionsSuccess,
-        });
-
-        failureData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${targetReportID}`,
-            value: copiedActionsFailure,
-        });
-    }
-
-    return [optimisticData, successData, failureData, optimisticReportActionCopyIDs];
-}
-
-function getReportFromHoldRequestsOnyxData({
-    chatReport,
-    iouReport,
-    recipient,
-    policy,
-    createdTimestamp,
-    betas,
-    isApprovalFlow = false,
-}: {
-    chatReport: OnyxTypes.Report;
-    iouReport: OnyxEntry<OnyxTypes.Report>;
-    recipient: Participant;
-    policy: OnyxEntry<OnyxTypes.Policy>;
-    createdTimestamp?: string;
-    betas: OnyxEntry<OnyxTypes.Beta[]>;
-    isApprovalFlow?: boolean;
-}): {
-    optimisticHoldReportID: string;
-    optimisticHoldActionID: string;
-    optimisticCreatedReportForUnapprovedTransactionsActionID: string | undefined;
-    optimisticHoldReportExpenseActionIDs: OptimisticHoldReportExpenseActionID[];
-    optimisticReportActionCopyIDs: OptimisticReportActionCopyIDs;
-    optimisticData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT | typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS | typeof ONYXKEYS.COLLECTION.TRANSACTION>>;
-    successData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>>;
-    failureData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT | typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS | typeof ONYXKEYS.COLLECTION.TRANSACTION>>;
-} {
-    const {holdReportActions, holdTransactions} = getHoldReportActionsAndTransactions(iouReport?.reportID);
-    const firstHoldTransaction = holdTransactions.at(0);
-    const newParentReportActionID = NumberUtils.rand64();
-
-    const coefficient = isExpenseReport(iouReport) ? -1 : 1;
-    const isPolicyExpenseChat = isPolicyExpenseChatReportUtil(chatReport);
-    const holdAmount = ((iouReport?.total ?? 0) - (iouReport?.unheldTotal ?? 0)) * coefficient;
-    const holdNonReimbursableAmount = ((iouReport?.nonReimbursableTotal ?? 0) - (iouReport?.unheldNonReimbursableTotal ?? 0)) * coefficient;
-
-    // Pass held transactions for formula computation (e.g., {report:startdate})
-    const reportTransactions: Record<string, OnyxTypes.Transaction> = {};
-    for (const holdTransaction of holdTransactions) {
-        if (holdTransaction?.transactionID) {
-            reportTransactions[holdTransaction.transactionID] = holdTransaction;
-        }
-    }
-
-    const optimisticExpenseReport = isPolicyExpenseChat
-        ? buildOptimisticExpenseReport({
-              chatReportID: chatReport.reportID,
-              policyID: chatReport.policyID ?? iouReport?.policyID,
-              payeeAccountID: recipient.accountID ?? 1,
-              total: holdAmount,
-              currency: iouReport?.currency ?? '',
-              nonReimbursableTotal: holdNonReimbursableAmount,
-              parentReportActionID: newParentReportActionID,
-              betas,
-              reportTransactions,
-              createdTimestamp,
-          })
-        : buildOptimisticIOUReport(
-              iouReport?.ownerAccountID ?? CONST.DEFAULT_NUMBER_ID,
-              iouReport?.managerID ?? CONST.DEFAULT_NUMBER_ID,
-              holdAmount,
-              chatReport.reportID,
-              iouReport?.currency ?? '',
-              false,
-              newParentReportActionID,
-              undefined,
-              createdTimestamp,
-          );
-
-    const optimisticExpenseReportPreview = buildOptimisticReportPreview(
-        chatReport,
-        optimisticExpenseReport,
-        '',
-        firstHoldTransaction,
-        optimisticExpenseReport.reportID,
-        newParentReportActionID,
-    );
-
-    let optimisticCreatedReportForUnapprovedAction: ReportAction | null = null;
-    if (isApprovalFlow) {
-        optimisticCreatedReportForUnapprovedAction = buildOptimisticCreatedReportForUnapprovedAction(optimisticExpenseReport.reportID, iouReport?.reportID);
-    }
-
-    const updateHeldReports: Record<string, Pick<OnyxTypes.Report, 'parentReportActionID' | 'parentReportID' | 'chatReportID'>> = {};
-    const addHoldReportActions: OnyxTypes.ReportActions = {};
-    const addHoldReportActionsSuccess: OnyxCollection<NullishDeep<ReportAction>> = {};
-    const deleteHoldReportActions: Record<string, Pick<OnyxTypes.ReportAction, 'message'>> = {};
-    const optimisticHoldReportExpenseActionIDs: OptimisticHoldReportExpenseActionID[] = [];
-
-    for (const holdReportAction of holdReportActions) {
-        // eslint-disable-next-line @typescript-eslint/non-nullable-type-assertion-style
-        const originalMessage = getOriginalMessage(holdReportAction) as OnyxTypes.OriginalMessageIOU;
-
-        deleteHoldReportActions[holdReportAction.reportActionID] = {
-            message: [
-                {
-                    deleted: DateUtils.getDBTime(),
-                    type: CONST.REPORT.MESSAGE.TYPE.TEXT,
-                    text: '',
-                },
-            ],
-        };
-
-        const reportActionID = NumberUtils.rand64();
-        addHoldReportActions[reportActionID] = {
-            ...holdReportAction,
-            reportActionID,
-            originalMessage: {
-                ...originalMessage,
-                IOUReportID: optimisticExpenseReport.reportID,
-            },
-            pendingAction: CONST.RED_BRICK_ROAD_PENDING_ACTION.ADD,
-        };
-        addHoldReportActionsSuccess[reportActionID] = {
-            pendingAction: null,
-        };
-
-        optimisticHoldReportExpenseActionIDs.push({optimisticReportActionID: reportActionID, oldReportActionID: holdReportAction.reportActionID});
-
-        const heldReport = getReportOrDraftReport(holdReportAction.childReportID);
-        if (heldReport) {
-            updateHeldReports[`${ONYXKEYS.COLLECTION.REPORT}${heldReport.reportID}`] = {
-                parentReportActionID: reportActionID,
-                parentReportID: optimisticExpenseReport.reportID,
-                chatReportID: optimisticExpenseReport.reportID,
-            };
-        }
-    }
-
-    const updateHeldTransactions: Record<string, Pick<OnyxTypes.Transaction, 'reportID'>> = {};
-    for (const transaction of holdTransactions) {
-        updateHeldTransactions[`${ONYXKEYS.COLLECTION.TRANSACTION}${transaction.transactionID}`] = {
-            reportID: optimisticExpenseReport.reportID,
-        };
-    }
-
-    const isApprovalEnabled = policy ? policy.approvalMode && policy.approvalMode !== CONST.POLICY.APPROVAL_MODE.OPTIONAL : false;
-
-    const optimisticData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT | typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS | typeof ONYXKEYS.COLLECTION.TRANSACTION>> = [
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT}${chatReport.reportID}`,
-            value: {
-                iouReportID: optimisticExpenseReport.reportID,
-                lastVisibleActionCreated: optimisticExpenseReportPreview.created,
-            },
-        },
-        // add new optimistic expense report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT}${optimisticExpenseReport.reportID}`,
-            value: {
-                ...optimisticExpenseReport,
-                unheldTotal: 0,
-                unheldNonReimbursableTotal: 0,
-                ...(isProcessingReport(iouReport) && isApprovalEnabled
-                    ? {
-                          stateNum: CONST.REPORT.STATE_NUM.SUBMITTED,
-                          statusNum: CONST.REPORT.STATUS_NUM.SUBMITTED,
-                      }
-                    : {}),
-            },
-        },
-        // add preview report action to main chat
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${chatReport.reportID}`,
-            value: {
-                [optimisticExpenseReportPreview.reportActionID]: optimisticExpenseReportPreview,
-            },
-        },
-        // remove hold report actions from old iou report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${iouReport?.reportID}`,
-            value: deleteHoldReportActions,
-        },
-        // add hold report actions to new iou report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: addHoldReportActions,
-        },
-        // update held reports with new parentReportActionID
-        {
-            onyxMethod: Onyx.METHOD.MERGE_COLLECTION,
-            key: `${ONYXKEYS.COLLECTION.REPORT}`,
-            value: updateHeldReports,
-        },
-        // update transactions with new iouReportID
-        {
-            onyxMethod: Onyx.METHOD.MERGE_COLLECTION,
-            key: `${ONYXKEYS.COLLECTION.TRANSACTION}`,
-            value: updateHeldTransactions,
-        },
-    ];
-
-    const bringReportActionsBack: Record<string, OnyxTypes.ReportAction> = {};
-    for (const reportAction of holdReportActions) {
-        bringReportActionsBack[reportAction.reportActionID] = reportAction;
-    }
-
-    const bringHeldTransactionsBack: Record<string, OnyxTypes.Transaction> = {};
-    for (const transaction of holdTransactions) {
-        bringHeldTransactionsBack[`${ONYXKEYS.COLLECTION.TRANSACTION}${transaction.transactionID}`] = transaction;
-    }
-
-    const successData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS>> = [
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${chatReport.reportID}`,
-            value: {
-                [optimisticExpenseReportPreview.reportActionID]: {
-                    pendingAction: null,
-                },
-            },
-        },
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: addHoldReportActionsSuccess,
-        },
-    ];
-
-    const failureData: Array<OnyxUpdate<typeof ONYXKEYS.COLLECTION.REPORT | typeof ONYXKEYS.COLLECTION.REPORT_ACTIONS | typeof ONYXKEYS.COLLECTION.TRANSACTION>> = [
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT}${chatReport.reportID}`,
-            value: {
-                iouReportID: chatReport.iouReportID,
-                lastVisibleActionCreated: chatReport.lastVisibleActionCreated,
-            },
-        },
-        // remove added optimistic expense report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT}${optimisticExpenseReport.reportID}`,
-            value: null,
-        },
-        // remove preview report action from the main chat
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${chatReport.reportID}`,
-            value: {
-                [optimisticExpenseReportPreview.reportActionID]: null,
-            },
-        },
-        // add hold report actions back to old iou report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${iouReport?.reportID}`,
-            value: bringReportActionsBack,
-        },
-        // remove hold report actions from the new iou report
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: null,
-        },
-        // add hold transactions back to old iou report
-        {
-            onyxMethod: Onyx.METHOD.MERGE_COLLECTION,
-            key: `${ONYXKEYS.COLLECTION.TRANSACTION}`,
-            value: bringHeldTransactionsBack,
-        },
-    ];
-
-    // Copy submission/approval actions to the new report
-    const [copiedActionsOptimistic, copiedActionsSuccess, copiedActionsFailure, optimisticReportActionCopyIDs] = getDuplicateActionsForPartialReport(
-        iouReport?.reportID,
-        optimisticExpenseReport.reportID,
-    );
-    // Only copy the report action for approval flow
-    if (isApprovalFlow && !isEmptyObject(optimisticReportActionCopyIDs)) {
-        optimisticData.push(...copiedActionsOptimistic);
-        successData.push(...copiedActionsSuccess);
-        failureData.push(...copiedActionsFailure);
-    }
-    // add optimistic system message explaining the created report for unapproved transactions
-    if (isApprovalFlow && optimisticCreatedReportForUnapprovedAction) {
-        optimisticData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: {
-                [optimisticCreatedReportForUnapprovedAction.reportActionID]: optimisticCreatedReportForUnapprovedAction,
-            },
-        });
-
-        successData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: {
-                [optimisticCreatedReportForUnapprovedAction.reportActionID]: {
-                    pendingAction: null,
-                    isOptimisticAction: null,
-                },
-            },
-        });
-
-        failureData.push({
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${optimisticExpenseReport.reportID}`,
-            value: {
-                [optimisticCreatedReportForUnapprovedAction.reportActionID]: null,
-            },
-        });
-    }
-
-    return {
-        optimisticData,
-        optimisticHoldActionID: optimisticExpenseReportPreview.reportActionID,
-        optimisticCreatedReportForUnapprovedTransactionsActionID: optimisticCreatedReportForUnapprovedAction?.reportActionID,
-        failureData,
-        successData,
-        optimisticHoldReportID: optimisticExpenseReport.reportID,
-        optimisticHoldReportExpenseActionIDs,
-        optimisticReportActionCopyIDs,
-    };
-}
-
 /**
  * Finds the participants for an IOU based on the attached report
  * @param transactionID of the transaction to set the participants of
@@ -5098,7 +4664,6 @@ export {
     setMoneyRequestReimbursable,
     startDistanceRequest,
     hasOutstandingChildRequest,
-    getReportFromHoldRequestsOnyxData,
     getUpdatedMoneyRequestReportData,
     getUpdateMoneyRequestParams,
     getUpdateTrackExpenseParams,
